requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.7
django-admin-interface==0.29.4
django-appconf==1.1.0
django-ckeditor-5==0.2.18
django-colorfield==0.12.0
django-cors-headers==4.7.0
django-filter==25.1
django-imagekit==5.0.0
django-quill-editor==0.1.42
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
frozenlist==1.6.2
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
multidict==6.4.4
packaging==24.2
pilkit==3.0
pillow==11.1.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.9.0
yarl==1.20.0
